fix: 放弃引入googletest，修复了多项式n输入非数时崩溃的问题，添加了x^-1特殊求导的错误文案，多项式现在支持更好的小数输入与显示

+++ src/TUI.h
@@ -113,9 +113,9 @@ int Page_Multinomial() {
 				Dialog_Info("第一个多项式如下：", { multinomials[0].print() });
 				break;
 			case 2:
-				int x;
+				double x;
 				// Dialog_Input("输入x = ：") >> x;
-				x = atoi(Dialog_Input("请输入x = ：").c_str());
+				x = atof(Dialog_Input("请输入x = ：").c_str());
 				Dialog_Info(" x = " + to_string(x) + "时，" + multinomials[0].print() + " =", { to_string(multinomials[0].calculate(x)) });
 				break;
 			case 3:
